pyproject.toml
[tool.poetry]
name = "ampel-ztf-archive"
version = "0.8.0-alpha.0"
description = "ZTF alert archive for the Ampel system"
authors = [
    "Jakob van Santen <jakob.van.santen@desy.de>",
]
maintainers = [
    "Jakob van Santen <jakob.van.santen@desy.de>"
]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://ampelproject.github.io"
repository = "https://github.com/AmpelProject/Ampel-ZTF-archive"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Typing :: Typed",
]
packages = [
    {include = "ampel"}
]

[tool.poetry.scripts]
ampel-ztf-archive-consumer-groups = 'ampel.ztf.archive.ArchiveDB:consumer_groups_command'

[tool.poetry.dependencies]
python = "^3.8"
sqlalchemy = "~1.3"
psycopg2 = "^2.8.6"
fastapi = {version = "^0.74.0", optional = true}
PyJWT = {version = "^2.1.0", extras = ["server"]}
boto3 = "^1.19.2"
fastavro = "^1.9.2"
astropy-healpix = "^0.7"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
coverage = {extras = ["toml"], version = "^6.3.1"}
mypy = "^1.10"
sqlalchemy-stubs = "^0.4"
pytest-timeout = "^1.4.2"
httpx = "^0.17.0"
pytest-mock = "^3.5.1"
pytest-asyncio = "^0.14.0"
moto = "^2.2.11"
boto3-stubs = {extras = ["s3"], version = "^1.19.2"}

[tool.poetry.extras]
server = ["fastapi"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
relative_files = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--showlocals -r a -v --integration"

[tool.mypy]
namespace_packages = true
plugins = [
  "sqlmypy"
]
show_error_codes = true
packages = [
  "ampel"
]

[[tool.mypy.overrides]]
module = [
  "astropy.*",
  "astropy_healpix.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "I",
    "UP",
    "B",
    # "DTZ",
    # "T20",
    # "PT",
    # "RET",
    # "SLF",
    # "SIM",
    # # "ARG", # sometimes your arguments have to conform to an interface
    # # "ERA", # seems to prohibit all comments, that's bad
    # "PL",
    # "PERF",
    # "RUF",
]

[tool.ruff.lint.per-file-ignores]
"ampel/ztf/archive/server/app.py" = ["B008"]
"ampel/ztf/archive/server/tokens.py" = ["B008"]